+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.21.0](https://github.com/actuate-rs/actuate/compare/actuate-v0.20.1...actuate-v0.21.0) - 2025-12-18
+
+## Breaking changes
+
+- Update to Bevy v0.17.3
+
 ## [0.20.1](https://github.com/actuate-rs/actuate/compare/actuate-v0.20.0...actuate-v0.20.1) - 2024-12-13
 
 ## Documentation
@@ -43,7 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Features
 
 - Add `use_effect` hook (5ae0a51)
-  - `fn use_effect<D, T>(cx: ScopeState, dependency: D, effect: impl FnOnce(&D))`
+  - `fn use_effect<D, T>(cx: ScopeState<'_>, dependency: D, effect: impl FnOnce(&D))`
 
 ## Fixes
 
@@ -205,7 +211,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - This allows for use of the `Data` macro without importing the full `prelude`.
 - Replace `DynCompose::new` with `dyn_compose` constructor fn (9d65ec8).
 - Return `Rc` from use_context
-  - `fn use_context<T: 'static>(cx: ScopeState) -> Result<&Rc<T>, ContextError<T>> { .. }`
+  - `fn use_context<T: 'static>(cx: ScopeState<'_>) -> Result<&Rc<T>, ContextError<T>> { .. }`
   - This allows for cloning context into `'static` environments.
 
 ### Refactors